Updates the "with-redux-saga" example
