+++ README.md
@@ -2,7 +2,7 @@
 
 Collection of usage examples of [Mock Service Worker](https://github.com/mswjs/msw) library.
 
-## Complete examples
+## Full examples
 
 > Complete examples showcase Mock Service Worker integration on all levels: for development, unit, and E2E tests.
 
@@ -13,4 +13,5 @@ Collection of usage examples of [Mock Service Worker](https://github.com/mswjs/m
 
 > Scoped examples illustrate the usage of Mock Service Worker with particular libraries or frameworks.
 
+- [Usage with **Redux Saga**](examples/with-redux-saga)
 - [Usage with **Storybook**](examples/with-storybook)
