+++ examples/with-redux-saga/package.json
