+++ examples/with-redux-saga/public/favicon.ico
