+++ examples/with-redux-saga/public/manifest.json
