+++ examples/with-redux-saga/public/mockServiceWorker.js
