+++ examples/with-redux-saga/src/index.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import {Provider} from 'react-redux'
-import {LoginForm} from './LoginForm'
-import {store} from './store'
+import { Provider } from 'react-redux'
+import { LoginForm } from './LoginForm'
+import { store } from './store'
 
 // Start the mocking conditionally.
 if (process.env.NODE_ENV === 'development') {
-  const {worker} = require('./mocks/browser')
+  const { worker } = require('./mocks/browser')
   worker.start()
 }
 
@@ -14,5 +14,5 @@ ReactDOM.render(
   <Provider store={store}>
     <LoginForm />
   </Provider>,
-  document.getElementById('root'),
+  document.getElementById('root')
 )
