+++ examples/with-redux-saga/src/mocks/browser.js
